fix: 修复资源泄漏和 API 一致性问题
P0 修复:
- CamoufoxFetcher: fetch() 使用 try/finally 确保 page.close()
- CamoufoxFetcher: _ensure_browser() 失败时清理 _context_manager

P1 修复:
- BrowserFetcher: timeout 转换为 int（避免 float 类型隐患）
- fetch()/post(): method 参数正确传递到底层 fetcher

P2 新增测试:
- test_fetch_get: GET 请求
- test_fetch_post: POST 请求（json 参数）
- test_post_method: post() 快捷方法
- test_batch_with_errors: 错误详情返回
- test_batch_with_errors_all_success: 全部成功场景
- test_max_concurrent: 并发参数验证

测试: 65 → 71 个，全部通过

+++ omnicrawl/fetchers/browser_fetcher.py
@@ -72,10 +72,10 @@ class BrowserFetcher(BaseFetcher):
         context = await self._browser.new_context(**context_kwargs)
         try:
             page = await context.new_page()
-            resp = await page.goto(url, wait_until="domcontentloaded", timeout=timeout * 1000)
+            resp = await page.goto(url, wait_until="domcontentloaded", timeout=int(timeout * 1000))
 
             if wait_for:
-                await page.wait_for_selector(wait_for, timeout=timeout * 1000)
+                await page.wait_for_selector(wait_for, timeout=int(timeout * 1000))
 
             html = await page.content()
             elapsed = time.time() - start
